update to ARGB

+++ lib/texture_tools.h
@@ -1,4 +1,4 @@
 #include "common.h"
 
-GLuint createTexture(std::vector<uint8_t>& framebuffer);
-void updateTexture(std::vector<uint8_t>& framebuffer ,GLuint tex);
+GLuint createTexture(std::vector<uint32_t>& framebuffer);
+void updateTexture(std::vector<uint32_t>& framebuffer ,GLuint tex);
